doge: Initialise dogecoin fork
We are trying to keep the dogecoin patchset as small as possible to
ease the maintenance burden.

The primary reason we need this fork is because in the Vault system we
want to be able to use _both_ `rust-bitcoin` and `rust-dogecoin`.

It would be nice to be able to use cargo's dependency aliasing feature
for this and carry this patchset as a branch on a `rust-bitcoin` tree
but cargo does not seem to allow depending on the same crate twice even
if its two different branches.

Rename package from `bitocoin` to `dogecoin`.

doge: Update readme